requirements.txt
azure-ai-textanalytics
streamlit
pandas
numpy
plotly
matplotlib
requests
python-dotenv
pyvis
wordcloud
altair
azure-ai-language-questionanswering
azure-ai-language-conversations
networkx
tqdm
azure-ai-vision-imageanalysis
torch>=1.10.0
transformers>=4.26.0
einops
pillow
timm
azure-ai-documentintelligence
pdf2image
Pillow
uuid
python-dateutil